feat: add recalculate button, change layout, adding token buttons WIP
